Improved landing page UI with modern healthcare design
