+++ apps/web/app/[locale]/about/page.tsx
@@ -2,7 +2,6 @@
 
 import {
     ShieldCheck,
-    Heart,
     Globe,
     MapPin,
     Mic,
@@ -23,7 +22,7 @@ export default function AboutPage() {
             {/* Hero */}
             <section className="border-b border-(--color-border-muted) bg-(--color-surface-page)">
                 <div className="container mx-auto max-w-6xl space-y-6 px-4 py-16 text-center md:py-24">
-                    <div className="inline-flex items-center gap-2 rounded-full border border-emerald-100 dark:border-emerald-900/30 bg-emerald-50 dark:bg-emerald-950/20 px-4 py-2 text-sm font-bold text-emerald-700 dark:text-emerald-400">
+                    <div className="inline-flex items-center gap-2 rounded-full border border-emerald-100 bg-emerald-50 px-4 py-2 text-sm font-bold text-emerald-700 dark:border-emerald-900/30 dark:bg-emerald-950/20 dark:text-emerald-400">
                         <span className="relative flex h-2 w-2">
                             <span className="absolute inline-flex h-full w-full animate-ping rounded-full bg-emerald-400 opacity-75"></span>
                             <span className="relative inline-flex h-2 w-2 rounded-full bg-emerald-500"></span>
@@ -31,7 +30,8 @@ export default function AboutPage() {
                         GSSoC 2026 Open Source Project
                     </div>
                     <h1 className="text-4xl leading-[1.1] font-black tracking-tight text-(--color-text-primary) md:text-6xl">
-                        About <span className="text-emerald-600 dark:text-emerald-400">SahiDawa</span>
+                        About{" "}
+                        <span className="text-emerald-600 dark:text-emerald-400">SahiDawa</span>
                     </h1>
                     <p className="mx-auto max-w-2xl text-lg leading-relaxed font-medium text-(--color-text-secondary) md:text-xl">
                         India's first open-source citizen medicine verifier & rural health bridge.
@@ -44,7 +44,7 @@ export default function AboutPage() {
                         <span className="flex items-center gap-1.5 rounded-full bg-(--color-surface-muted) px-3 py-1.5 text-sm font-semibold text-(--color-text-secondary)">
                             <Globe size={14} /> 22 Indian Languages
                         </span>
-                        <span className="flex items-center gap-1.5 rounded-full bg-emerald-100 dark:bg-emerald-950/30 px-3 py-1.5 text-sm font-semibold text-emerald-700 dark:text-emerald-400">
+                        <span className="flex items-center gap-1.5 rounded-full bg-emerald-100 px-3 py-1.5 text-sm font-semibold text-emerald-700 dark:bg-emerald-950/30 dark:text-emerald-400">
                             <Star size={14} /> Open Source MIT License
                         </span>
                     </div>
@@ -90,7 +90,9 @@ export default function AboutPage() {
                             <div className="mb-4 flex h-14 w-14 items-center justify-center rounded-2xl bg-(--color-surface-muted) text-(--color-text-secondary) transition-all duration-300 group-hover:bg-emerald-500 group-hover:text-white">
                                 {item.icon}
                             </div>
-                            <h3 className="mb-2 text-lg font-bold text-(--color-text-primary)">{item.title}</h3>
+                            <h3 className="mb-2 text-lg font-bold text-(--color-text-primary)">
+                                {item.title}
+                            </h3>
                             <p className="text-sm leading-relaxed font-medium text-(--color-text-secondary)">
                                 {item.desc}
                             </p>
@@ -99,9 +101,9 @@ export default function AboutPage() {
                 </div>
 
                 {/* Incident box */}
-                <div className="mt-8 rounded-3xl border border-orange-200 dark:border-orange-900/30 bg-orange-50 dark:bg-orange-950/20 p-6 md:p-8">
+                <div className="mt-8 rounded-3xl border border-orange-200 bg-orange-50 p-6 md:p-8 dark:border-orange-900/30 dark:bg-orange-950/20">
                     <div className="flex items-start gap-4">
-                        <div className="flex h-10 w-10 shrink-0 items-center justify-center rounded-full bg-orange-100 dark:bg-orange-950/40 text-orange-500 dark:text-orange-400">
+                        <div className="flex h-10 w-10 shrink-0 items-center justify-center rounded-full bg-orange-100 text-orange-500 dark:bg-orange-950/40 dark:text-orange-400">
                             <AlertTriangle size={20} strokeWidth={2.5} />
                         </div>
                         <div>
@@ -129,11 +131,13 @@ export default function AboutPage() {
             <section className="border-y border-(--color-border-muted) bg-(--color-surface-page)">
                 <div className="container mx-auto max-w-6xl px-4 py-16">
                     <div className="grid grid-cols-1 gap-8 md:grid-cols-2">
-                        <div className="rounded-3xl border border-emerald-100 dark:border-emerald-900/30 bg-emerald-50 dark:bg-emerald-950/20 p-8">
+                        <div className="rounded-3xl border border-emerald-100 bg-emerald-50 p-8 dark:border-emerald-900/30 dark:bg-emerald-950/20">
                             <div className="mb-6 flex h-14 w-14 items-center justify-center rounded-2xl bg-emerald-500 text-white shadow-lg shadow-emerald-500/25">
                                 <ShieldCheck size={28} strokeWidth={2.5} />
                             </div>
-                            <h3 className="mb-3 text-2xl font-black text-(--color-text-primary)">Our Mission</h3>
+                            <h3 className="mb-3 text-2xl font-black text-(--color-text-primary)">
+                                Our Mission
+                            </h3>
                             <p className="leading-relaxed font-medium text-(--color-text-secondary)">
                                 To empower every Indian citizen — regardless of language, location,
                                 or literacy — with the ability to instantly verify medicines, access
@@ -165,18 +169,38 @@ export default function AboutPage() {
                 </div>
                 <div className="grid grid-cols-2 gap-6 text-center md:grid-cols-4">
                     {[
-                        { icon: <Unlock className="h-10 w-10 text-emerald-500" />, title: "Open Source", desc: "MIT Licensed. Always." },
-                        { icon: <BadgeCheck className="h-10 w-10 text-blue-500" />, title: "Free Forever", desc: "No hidden costs. Ever." },
-                        { icon: <Globe className="h-10 w-10 text-purple-500" />, title: "Inclusive", desc: "22 languages. All of Bharat." },
-                        { icon: <Lock className="h-10 w-10 text-slate-500" />, title: "Privacy First", desc: "No data sold. No ads." },
+                        {
+                            icon: <Unlock className="h-10 w-10 text-emerald-500" />,
+                            title: "Open Source",
+                            desc: "MIT Licensed. Always.",
+                        },
+                        {
+                            icon: <BadgeCheck className="h-10 w-10 text-blue-500" />,
+                            title: "Free Forever",
+                            desc: "No hidden costs. Ever.",
+                        },
+                        {
+                            icon: <Globe className="h-10 w-10 text-purple-500" />,
+                            title: "Inclusive",
+                            desc: "22 languages. All of Bharat.",
+                        },
+                        {
+                            icon: <Lock className="h-10 w-10 text-slate-500" />,
+                            title: "Privacy First",
+                            desc: "No data sold. No ads.",
+                        },
                     ].map((v, i) => (
                         <div
                             key={i}
                             className="rounded-3xl border border-(--color-border-muted) bg-(--color-surface-page) p-6"
                         >
                             <div className="mb-3 flex justify-center">{v.icon}</div>
-                            <h3 className="mb-1 font-bold text-(--color-text-primary)">{v.title}</h3>
-                            <p className="text-sm font-medium text-(--color-text-secondary)">{v.desc}</p>
+                            <h3 className="mb-1 font-bold text-(--color-text-primary)">
+                                {v.title}
+                            </h3>
+                            <p className="text-sm font-medium text-(--color-text-secondary)">
+                                {v.desc}
+                            </p>
                         </div>
                     ))}
                 </div>
@@ -200,7 +224,7 @@ export default function AboutPage() {
                         </p>
                         <div className="pt-2">
                             <Link href="/contact">
-                                <button className="inline-flex items-center gap-2 rounded-2xl bg-white dark:bg-slate-900 px-8 py-3.5 text-base font-bold text-emerald-600 dark:text-emerald-400 shadow-lg transition-all duration-200 hover:scale-105 hover:shadow-xl">
+                                <button className="inline-flex items-center gap-2 rounded-2xl bg-white px-8 py-3.5 text-base font-bold text-emerald-600 shadow-lg transition-all duration-200 hover:scale-105 hover:shadow-xl dark:bg-slate-900 dark:text-emerald-400">
                                     Contact Us <ChevronRight size={18} />
                                 </button>
                             </Link>
