+++ apps/web/app/[locale]/components/PageHeader.tsx
@@ -80,7 +80,7 @@ export const PageHeader = ({
                         </div>
                     ) : isDark ? (
                         <button
-                            onClick={() => console.log("Quick actions menu triggered!")}
+                            onClick={() => {}}
                             aria-label="Quick actions"
                             className={`flex h-10 w-10 items-center justify-center rounded-full bg-white/10 backdrop-blur-md transition-colors hover:bg-white/20 ${pageHeaderFocusRingClass}`}
                         >
