+++ apps/web/app/[locale]/globals.css
@@ -46,14 +46,6 @@
     --color-dark-scrollbar-hover: #64748b;
     --color-scrollbar-thumb: #cbd5e1;
 }
-:root {
-    --color-surface-page: #ffffff;
-    --color-text-primary: #1e293b;
-    /* ... your other light mode variables ... */
-}
-
-/* ADD THIS ENTIRE BLOCK: 
-   When the dark class is active, swap the variables to dark colors! */
 .dark {
     --color-surface-page: #0f172a;
     --color-surface-muted: #1e293b;
@@ -242,7 +234,177 @@
 }
 
 @theme {
-    --font-sans: var(--font-geist-sans), ui-sans-serif, system-ui, sans-serif;
+    --font-sans: var(--font-instrument), ui-sans-serif, system-ui, sans-serif;
+}
+
+*, *::before, *::after { box-sizing: border-box; }
+
+html { scroll-behavior: smooth; }
+
+body { font-family: var(--font-instrument), 'Instrument Sans', sans-serif; }
+
+.display, h1, h2, h3, h4 {
+    font-family: var(--font-bricolage), 'Bricolage Grotesque', sans-serif;
+    letter-spacing: -0.025em;
+}
+
+@keyframes fade-up {
+    from { opacity: 0; transform: translateY(28px); }
+    to   { opacity: 1; transform: translateY(0); }
+}
+@keyframes fade-in {
+    from { opacity: 0; }
+    to   { opacity: 1; }
+}
+@keyframes float-slow {
+    0%, 100% { transform: translateY(0px) rotate(0deg); }
+    33%       { transform: translateY(-10px) rotate(1deg); }
+    66%       { transform: translateY(-5px) rotate(-1deg); }
+}
+@keyframes float-med {
+    0%, 100% { transform: translateY(0px); }
+    50%       { transform: translateY(-14px); }
+}
+@keyframes spin-slow {
+    from { transform: rotate(0deg); }
+    to   { transform: rotate(360deg); }
+}
+@keyframes scan-laser {
+    0%   { top: 12%; }
+    50%  { top: 82%; }
+    100% { top: 12%; }
+}
+@keyframes shimmer {
+    0%   { background-position: -200% center; }
+    100% { background-position:  200% center; }
+}
+@keyframes orbit {
+    from { transform: rotate(0deg) translateX(120px) rotate(0deg); }
+    to   { transform: rotate(360deg) translateX(120px) rotate(-360deg); }
+}
+@keyframes counter-orbit {
+    from { transform: rotate(0deg) translateX(80px) rotate(0deg); }
+    to   { transform: rotate(-360deg) translateX(80px) rotate(360deg); }
+}
+@keyframes ping-slow {
+    0%   { transform: scale(1); opacity: 0.8; }
+    100% { transform: scale(2.2); opacity: 0; }
+}
+@keyframes bar-grow {
+    from { transform: scaleY(0); }
+    to   { transform: scaleY(1); }
+}
+@keyframes slide-right {
+    from { transform: translateX(-100%); }
+    to   { transform: translateX(0); }
+}
+@keyframes marquee {
+    from { transform: translateX(0); }
+    to   { transform: translateX(-50%); }
+}
+
+.anim-fade-up   { animation: fade-up  0.7s cubic-bezier(.22,1,.36,1) both; }
+.anim-fade-in   { animation: fade-in  0.6s ease both; }
+.anim-float     { animation: float-slow 7s ease-in-out infinite; }
+.anim-float-med { animation: float-med  5s ease-in-out infinite; }
+.anim-spin-slow { animation: spin-slow 22s linear infinite; }
+.anim-spin-rev  { animation: spin-slow 18s linear infinite reverse; }
+.anim-orbit     { animation: orbit 14s linear infinite; }
+.anim-c-orbit   { animation: counter-orbit 10s linear infinite; }
+.anim-marquee   { animation: marquee 28s linear infinite; }
+
+.delay-100 { animation-delay: 100ms; }
+.delay-200 { animation-delay: 200ms; }
+.delay-300 { animation-delay: 300ms; }
+.delay-400 { animation-delay: 400ms; }
+.delay-500 { animation-delay: 500ms; }
+.delay-600 { animation-delay: 600ms; }
+.delay-700 { animation-delay: 700ms; }
+.delay-800 { animation-delay: 800ms; }
+
+.scan-laser {
+    position: absolute; left: 0; right: 0; height: 2px;
+    background: linear-gradient(90deg, transparent 0%, #2dd4bf 20%, #67e8f9 50%, #2dd4bf 80%, transparent 100%);
+    box-shadow: 0 0 18px 4px rgba(45,212,191,0.7);
+    animation: scan-laser 2.8s ease-in-out infinite;
+}
+
+.shimmer-text {
+    background: linear-gradient(90deg, #0d9488 0%, #06b6d4 30%, #38bdf8 50%, #06b6d4 70%, #0d9488 100%);
+    background-size: 200% auto;
+    -webkit-background-clip: text;
+    background-clip: text;
+    -webkit-text-fill-color: transparent;
+    animation: shimmer 4s linear infinite;
+}
+
+.glass {
+    background: rgba(255,255,255,0.72);
+    backdrop-filter: blur(20px) saturate(180%);
+    -webkit-backdrop-filter: blur(20px) saturate(180%);
+    border: 1px solid rgba(255,255,255,0.55);
+}
+
+.glass-dark {
+    background: rgba(10,15,30,0.65);
+    backdrop-filter: blur(20px) saturate(160%);
+    -webkit-backdrop-filter: blur(20px) saturate(160%);
+    border: 1px solid rgba(255,255,255,0.08);
+}
+
+.card-lift {
+    transition: transform 0.35s cubic-bezier(.22,1,.36,1), box-shadow 0.35s cubic-bezier(.22,1,.36,1);
+}
+.card-lift:hover {
+    transform: translateY(-6px);
+    box-shadow: 0 28px 60px -12px rgba(13,148,136,0.18);
+}
+
+.btn-primary {
+    position: relative; overflow: hidden;
+}
+.btn-primary::before {
+    content: '';
+    position: absolute; inset: 0;
+    background: linear-gradient(135deg, rgba(255,255,255,0.18) 0%, transparent 60%);
+    opacity: 0;
+    transition: opacity 0.3s;
+}
+.btn-primary:hover::before { opacity: 1; }
+
+.underline-wave {
+    position: relative; display: inline-block;
+}
+.underline-wave::after {
+    content: '';
+    position: absolute;
+    bottom: -4px; left: 0; right: 0; height: 3px;
+    background: linear-gradient(90deg, #0d9488, #06b6d4, #38bdf8);
+    border-radius: 999px;
+    transform: scaleX(0);
+    transform-origin: left;
+    transition: transform 0.4s cubic-bezier(.22,1,.36,1);
+}
+.underline-wave:hover::after { transform: scaleX(1); }
+
+.ticker-wrap { overflow: hidden; }
+.ticker { display: flex; width: max-content; }
+
+.step-connector {
+    position: absolute;
+    top: 40px;
+    height: 1px;
+}
+
+/* Noise texture overlay */
+.noise::after {
+    content: '';
+    position: absolute;
+    inset: 0;
+    background-image: url("data:image/svg+xml,%3Csvg viewBox='0 0 256 256' xmlns='http://www.w3.org/2000/svg'%3E%3Cfilter id='noise'%3E%3CfeTurbulence type='fractalNoise' baseFrequency='0.9' numOctaves='4' stitchTiles='stitch'/%3E%3C/filter%3E%3Crect width='100%25' height='100%25' filter='url(%23noise)' opacity='1'/%3E%3C/svg%3E");
+    opacity: 0.03;
+    pointer-events: none;
+    border-radius: inherit;
 }
 
 /* Custom scrollbar for chat */
