+++ apps/web/app/[locale]/login/page.tsx
@@ -40,7 +40,7 @@ export default function LoginPage() {
 
                 router.push("/reports/me");
             }
-        } catch (err) {
+        } catch {
             setError("Something went wrong. Please try again.");
         }
 
@@ -52,13 +52,15 @@ export default function LoginPage() {
             <div className="w-full max-w-md">
                 {/* Logo */}
                 <div className="mb-8 flex items-center justify-center gap-3">
-                    <div className="rounded-2xl bg-emerald-100 dark:bg-emerald-950/30 p-3 shadow-sm">
-                        <ShieldCheck className="h-7 w-7 text-emerald-600 dark:text-emerald-450" />
+                    <div className="rounded-2xl bg-emerald-100 p-3 shadow-sm dark:bg-emerald-950/30">
+                        <ShieldCheck className="dark:text-emerald-450 h-7 w-7 text-emerald-600" />
                     </div>
 
                     <div>
                         <h1 className="text-3xl font-bold text-(--color-text-primary)">SahiDawa</h1>
-                        <p className="text-sm text-(--color-text-secondary)">Secure Health Verification</p>
+                        <p className="text-sm text-(--color-text-secondary)">
+                            Secure Health Verification
+                        </p>
                     </div>
                 </div>
 
@@ -66,7 +68,7 @@ export default function LoginPage() {
                 <div className="rounded-3xl border border-(--color-border-muted) bg-(--color-surface-page) p-8 shadow-xl">
                     <div className="mb-7">
                         <h2 className="flex items-center gap-2 text-3xl font-bold text-(--color-text-primary)">
-                            Welcome Back <Hand className="h-8 w-8 text-amber-500 animate-bounce" />
+                            Welcome Back <Hand className="h-8 w-8 animate-bounce text-amber-500" />
                         </h2>
 
                         <p className="mt-2 text-(--color-text-secondary)">
@@ -78,7 +80,7 @@ export default function LoginPage() {
                     {error && (
                         <LiveMessage
                             tone="critical"
-                            className="mb-5 rounded-xl border border-red-200 dark:border-red-900 bg-red-50 dark:bg-red-950/20 px-4 py-3 text-sm text-red-600 dark:text-red-400"
+                            className="mb-5 rounded-xl border border-red-200 bg-red-50 px-4 py-3 text-sm text-red-600 dark:border-red-900 dark:bg-red-950/20 dark:text-red-400"
                         >
                             {error}
                         </LiveMessage>
@@ -107,7 +109,9 @@ export default function LoginPage() {
 
                         {/* Password */}
                         <div>
-                            <label className="text-sm font-medium text-(--color-text-primary)">Password</label>
+                            <label className="text-sm font-medium text-(--color-text-primary)">
+                                Password
+                            </label>
 
                             <div className="mt-2 flex items-center gap-3 rounded-2xl border border-(--color-border-muted) bg-(--color-surface-muted) px-4 py-3 transition focus-within:border-emerald-500 focus-within:bg-(--color-surface-page)">
                                 <Lock className="h-5 w-5 text-(--color-text-muted)" />
@@ -127,7 +131,7 @@ export default function LoginPage() {
                         <button
                             type="submit"
                             disabled={loading}
-                            className="mt-2 flex w-full items-center justify-center gap-2 rounded-2xl bg-emerald-600 py-3.5 font-semibold text-white shadow-lg shadow-emerald-250/20 dark:shadow-emerald-950/20 transition-all hover:bg-emerald-700"
+                            className="shadow-emerald-250/20 mt-2 flex w-full items-center justify-center gap-2 rounded-2xl bg-emerald-600 py-3.5 font-semibold text-white shadow-lg transition-all hover:bg-emerald-700 dark:shadow-emerald-950/20"
                         >
                             {loading ? "Signing In..." : "Sign In"}
