+++ apps/web/app/[locale]/reports/me/page.tsx
@@ -7,14 +7,12 @@ import {
     Clock,
     FileText,
     ImageOff,
-    Loader2,
     LogIn,
     MapPin,
     RefreshCw,
     ShieldCheck,
     XCircle,
 } from "lucide-react";
-import { Link } from "@/i18n/routing";
 import { PageHeader } from "../../components/PageHeader";
 import Card from "@/components/Card";
 import LazyImage from "@/components/LazyImage";
@@ -106,8 +104,8 @@ function ReportCard({ report }: { report: MyReport }) {
         report.reported_brand_name?.trim() || report.scanned_barcode || "Unnamed medicine";
 
     return (
-        <Card className="flex flex-col sm:flex-row bg-(--color-surface-page) border-(--color-border-muted) shadow-sm overflow-hidden">
-            <div className="flex h-40 shrink-0 items-center justify-center bg-(--color-surface-muted) sm:h-32 sm:w-32 border-r border-(--color-border-muted)">
+        <Card className="flex flex-col overflow-hidden border-(--color-border-muted) bg-(--color-surface-page) shadow-sm sm:flex-row">
+            <div className="flex h-40 shrink-0 items-center justify-center border-r border-(--color-border-muted) bg-(--color-surface-muted) sm:h-32 sm:w-32">
                 {isSafePhotoUrl(report.photo_url) ? (
                     <LazyImage
                         src={report.photo_url}
@@ -249,8 +247,11 @@ export default function MyReportsPage() {
                 {state.kind === "loading" && (
                     <div className="flex flex-col gap-3" aria-label="Loading your reports">
                         {[1, 2, 3].map((i) => (
-                            <Card key={i} className="flex flex-col sm:flex-row bg-(--color-surface-page) border-(--color-border-muted)">
-                                <Skeleton className="h-40 shrink-0 sm:h-32 sm:w-32 rounded-none bg-slate-200 dark:bg-slate-800" />
+                            <Card
+                                key={i}
+                                className="flex flex-col border-(--color-border-muted) bg-(--color-surface-page) sm:flex-row"
+                            >
+                                <Skeleton className="h-40 shrink-0 rounded-none bg-slate-200 sm:h-32 sm:w-32 dark:bg-slate-800" />
                                 <div className="flex min-w-0 flex-1 flex-col gap-2 p-4">
                                     <div className="flex items-start justify-between gap-3">
                                         <Skeleton className="h-5 w-1/2 bg-slate-200 dark:bg-slate-800" />
@@ -284,7 +285,7 @@ export default function MyReportsPage() {
                         description={state.message}
                         actionLabel="Try again"
                         onAction={fetchMine}
-                        className="border-rose-200 dark:border-rose-950/40 bg-(--color-surface-page)!"
+                        className="border-rose-200 bg-(--color-surface-page)! dark:border-rose-950/40"
                     />
                 )}
