+++ apps/web/app/[locale]/voice/VoicePanels.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-import { useId } from "react";
 import {
     AlertTriangle,
     Mic,
@@ -53,8 +52,12 @@ export function VoiceIntroPanel({
     return (
         <div className="animate-in fade-in slide-in-from-bottom-4 space-y-8 text-center duration-500">
             <div className="space-y-3">
-                <h1 className="text-4xl font-black tracking-tight text-(--color-text-primary)">{title}</h1>
-                <p className="mx-auto max-w-xs font-medium text-(--color-text-secondary)">{subtitle}</p>
+                <h1 className="text-4xl font-black tracking-tight text-(--color-text-primary)">
+                    {title}
+                </h1>
+                <p className="mx-auto max-w-xs font-medium text-(--color-text-secondary)">
+                    {subtitle}
+                </p>
             </div>
 
             <div className="mx-auto grid max-w-sm grid-cols-2 gap-4">
@@ -63,14 +66,22 @@ export function VoiceIntroPanel({
                     <p className="text-xs font-bold tracking-tighter text-(--color-text-muted) uppercase">
                         {exampleLabel}
                     </p>
-                    <p className="mt-1 text-sm font-bold text-(--color-text-primary)">{exampleText}</p>
+                    <p className="mt-1 text-sm font-bold text-(--color-text-primary)">
+                        {exampleText}
+                    </p>
                 </div>
                 <div className="rounded-3xl border border-(--color-border-muted) bg-(--color-surface-page) p-4 text-left shadow-sm">
-                    <Volume2 size={20} aria-hidden="true" className="mb-2 text-emerald-700 dark:text-emerald-400" />
+                    <Volume2
+                        size={20}
+                        aria-hidden="true"
+                        className="mb-2 text-emerald-700 dark:text-emerald-400"
+                    />
                     <p className="text-xs font-bold tracking-tighter text-(--color-text-muted) uppercase">
                         {assistantLabel}
                     </p>
-                    <p className="mt-1 text-sm font-bold text-(--color-text-primary)">{assistantValue}</p>
+                    <p className="mt-1 text-sm font-bold text-(--color-text-primary)">
+                        {assistantValue}
+                    </p>
                 </div>
             </div>
         </div>
@@ -124,11 +135,13 @@ export function VoiceListeningPanel({
             <p className="text-center text-2xl font-bold text-(--color-text-primary) italic">
                 {transcript || "…"}
             </p>
-            <p className="text-sm font-bold tracking-widest text-emerald-700 dark:text-emerald-400 uppercase">
+            <p className="text-sm font-bold tracking-widest text-emerald-700 uppercase dark:text-emerald-400">
                 {statusLabel}
             </p>
             {helperLabel ? (
-                <p className="text-center text-sm font-medium text-(--color-text-secondary)">{helperLabel}</p>
+                <p className="text-center text-sm font-medium text-(--color-text-secondary)">
+                    {helperLabel}
+                </p>
             ) : null}
         </div>
     );
@@ -145,7 +158,7 @@ export function VoiceProcessingPanel({ title, subtitle }: { title: string; subti
             <div className="relative" aria-hidden="true">
                 <div className="h-24 w-24 animate-spin rounded-full border-4 border-(--color-surface-muted) border-t-emerald-600 motion-reduce:animate-none"></div>
                 <Sparkles
-                    className="absolute inset-0 m-auto animate-pulse text-emerald-700 dark:text-emerald-400 motion-reduce:animate-none"
+                    className="absolute inset-0 m-auto animate-pulse text-emerald-700 motion-reduce:animate-none dark:text-emerald-400"
                     size={32}
                     aria-hidden="true"
                 />
@@ -210,7 +223,7 @@ export function VoiceReviewPanel({
     return (
         <div className="animate-in fade-in slide-in-from-bottom-8 w-full max-w-md rounded-[2.5rem] border border-(--color-border-muted) bg-(--color-surface-page) p-8 shadow-xl duration-500 motion-reduce:animate-none">
             {showEmergency && (
-                <div className="mb-6 rounded-2xl border border-red-200 dark:border-red-900 bg-red-50 dark:bg-red-950/20 p-4 text-left">
+                <div className="mb-6 rounded-2xl border border-red-200 bg-red-50 p-4 text-left dark:border-red-900 dark:bg-red-950/20">
                     <div className="flex items-start gap-3">
                         <AlertTriangle
                             className="mt-0.5 shrink-0 text-red-600 dark:text-red-400"
@@ -218,7 +231,9 @@ export function VoiceReviewPanel({
                             aria-hidden="true"
                         />
                         <div>
-                            <p className="font-bold text-red-900 dark:text-red-200">{emergencyTitle}</p>
+                            <p className="font-bold text-red-900 dark:text-red-200">
+                                {emergencyTitle}
+                            </p>
                             <p className="mt-1 text-sm leading-relaxed text-red-800 dark:text-red-400">
                                 {emergencyBody}
                             </p>
@@ -227,12 +242,14 @@ export function VoiceReviewPanel({
                 </div>
             )}
             <div className="mb-6 flex items-start gap-3">
-                <div className="flex h-12 w-12 shrink-0 items-center justify-center rounded-2xl bg-amber-50 dark:bg-amber-950/30 text-amber-600 dark:text-amber-450">
+                <div className="dark:text-amber-450 flex h-12 w-12 shrink-0 items-center justify-center rounded-2xl bg-amber-50 text-amber-600 dark:bg-amber-950/30">
                     <AlertTriangle size={24} aria-hidden="true" />
                 </div>
                 <div className="space-y-2">
                     <h2 className="font-black text-(--color-text-primary)">{title}</h2>
-                    <p className="text-sm leading-relaxed text-(--color-text-secondary)">{message}</p>
+                    <p className="text-sm leading-relaxed text-(--color-text-secondary)">
+                        {message}
+                    </p>
                     <ConfidenceBadge
                         confidence={confidence}
                         labelPrefix={confidenceLabelPrefix}
@@ -248,7 +265,7 @@ export function VoiceReviewPanel({
             <div className="mt-6 grid grid-cols-1 gap-3 sm:grid-cols-2">
                 <button
                     onClick={onRetry}
-                    className={`w-full rounded-2xl bg-(--color-surface-muted) py-3 font-bold text-(--color-text-primary) border border-(--color-border-muted) transition-colors hover:bg-(--color-border-muted) ${VOICE_FOCUS_RING_CLASS}`}
+                    className={`w-full rounded-2xl border border-(--color-border-muted) bg-(--color-surface-muted) py-3 font-bold text-(--color-text-primary) transition-colors hover:bg-(--color-border-muted) ${VOICE_FOCUS_RING_CLASS}`}
                 >
                     {retryLabel}
                 </button>
@@ -279,13 +296,13 @@ export function VoiceErrorPanel({
     const showTextFallback = error.type === "timeout" || error.type === "service-unavailable";
     return (
         <div
-            className="animate-in fade-in slide-in-from-bottom-8 w-full max-w-md rounded-[2.5rem] border border-red-100 dark:border-red-950/40 bg-(--color-surface-page) p-8 shadow-xl duration-500 motion-reduce:animate-none"
+            className="animate-in fade-in slide-in-from-bottom-8 w-full max-w-md rounded-[2.5rem] border border-red-100 bg-(--color-surface-page) p-8 shadow-xl duration-500 motion-reduce:animate-none dark:border-red-950/40"
             role="alert"
             aria-live="assertive"
             aria-describedby="voice-error-message"
         >
             <div className="mb-6 flex items-start gap-3">
-                <div className="flex h-12 w-12 shrink-0 items-center justify-center rounded-2xl bg-red-50 dark:bg-red-950/30 text-red-600 dark:text-red-400">
+                <div className="flex h-12 w-12 shrink-0 items-center justify-center rounded-2xl bg-red-50 text-red-600 dark:bg-red-950/30 dark:text-red-400">
                     <AlertTriangle size={24} aria-hidden="true" />
                 </div>
                 <div>
@@ -302,7 +319,7 @@ export function VoiceErrorPanel({
             <div className="flex flex-col gap-3">
                 <button
                     onClick={onRetry}
-                    className={`w-full rounded-2xl bg-slate-900 dark:bg-slate-100 dark:text-slate-900 dark:hover:bg-slate-200 py-4 font-bold text-white transition-all hover:bg-slate-800 ${VOICE_FOCUS_RING_CLASS}`}
+                    className={`w-full rounded-2xl bg-slate-900 py-4 font-bold text-white transition-all hover:bg-slate-800 dark:bg-slate-100 dark:text-slate-900 dark:hover:bg-slate-200 ${VOICE_FOCUS_RING_CLASS}`}
                 >
                     {retryLabel}
                 </button>
@@ -370,11 +387,14 @@ export function VoiceResultPanel({
             aria-labelledby="voice-ai-analysis-heading"
         >
             <div className="mb-6 flex items-center gap-3">
-                <div className="flex h-12 w-12 items-center justify-center rounded-2xl bg-blue-50 dark:bg-blue-950/30 text-blue-600 dark:text-blue-400">
+                <div className="flex h-12 w-12 items-center justify-center rounded-2xl bg-blue-50 text-blue-600 dark:bg-blue-950/30 dark:text-blue-400">
                     <Sparkles size={24} aria-hidden="true" />
                 </div>
                 <div>
-                    <h2 id="voice-ai-analysis-heading" className="font-black text-(--color-text-primary)">
+                    <h2
+                        id="voice-ai-analysis-heading"
+                        className="font-black text-(--color-text-primary)"
+                    >
                         {heading}
                     </h2>
                     <p className="text-xs font-bold tracking-tighter text-(--color-text-muted) uppercase">
@@ -385,7 +405,7 @@ export function VoiceResultPanel({
 
             <div className="space-y-4">
                 {result.emergency && (
-                    <div className="rounded-2xl border border-red-200 dark:border-red-900 bg-red-50 dark:bg-red-950/20 p-4 text-left">
+                    <div className="rounded-2xl border border-red-200 bg-red-50 p-4 text-left dark:border-red-900 dark:bg-red-950/20">
                         <div className="flex items-start gap-3">
                             <AlertTriangle
                                 className="mt-0.5 shrink-0 text-red-600 dark:text-red-400"
@@ -393,7 +413,9 @@ export function VoiceResultPanel({
                                 aria-hidden="true"
                             />
                             <div>
-                                <p className="font-bold text-red-900 dark:text-red-200">{emergencyTitle}</p>
+                                <p className="font-bold text-red-900 dark:text-red-200">
+                                    {emergencyTitle}
+                                </p>
                                 <p className="mt-1 text-sm leading-relaxed text-red-800 dark:text-red-400">
                                     {emergencyBody}
                                 </p>
@@ -413,11 +435,15 @@ export function VoiceResultPanel({
                             valueLabel={confidenceValueLabel}
                         />
                     </div>
-                    <p className="text-sm leading-relaxed text-(--color-text-primary)">{transcript}</p>
+                    <p className="text-sm leading-relaxed text-(--color-text-primary)">
+                        {transcript}
+                    </p>
                 </div>
 
                 <div className="rounded-2xl border border-(--color-border-muted) bg-(--color-surface-muted) p-4">
-                    <p className="text-sm leading-relaxed text-(--color-text-primary)">{result.summary}</p>
+                    <p className="text-sm leading-relaxed text-(--color-text-primary)">
+                        {result.summary}
+                    </p>
                 </div>
 
                 {result.recommendations.length > 0 && (
@@ -429,7 +455,7 @@ export function VoiceResultPanel({
                             {result.recommendations.map((recommendation, index) => (
                                 <div
                                     key={`${recommendation}-${index}`}
-                                    className="flex items-center gap-4 rounded-2xl border border-emerald-100 dark:border-emerald-900/30 bg-emerald-50 dark:bg-emerald-950/20 p-4"
+                                    className="flex items-center gap-4 rounded-2xl border border-emerald-100 bg-emerald-50 p-4 dark:border-emerald-900/30 dark:bg-emerald-950/20"
                                 >
                                     <div className="flex h-10 w-10 shrink-0 items-center justify-center rounded-full bg-emerald-600 text-white">
                                         <span className="font-bold">{index + 1}</span>
@@ -444,13 +470,15 @@ export function VoiceResultPanel({
                 )}
 
                 <div className="rounded-2xl border border-(--color-border-muted) bg-(--color-surface-muted) px-4 py-3">
-                    <p className="text-xs leading-relaxed text-(--color-text-secondary)">{result.disclaimer}</p>
+                    <p className="text-xs leading-relaxed text-(--color-text-secondary)">
+                        {result.disclaimer}
+                    </p>
                 </div>
 
                 <div className="grid grid-cols-1 gap-3 sm:grid-cols-2">
                     <button
                         onClick={onShare}
-                        className={`flex w-full items-center justify-center gap-2 rounded-2xl bg-(--color-surface-muted) py-3 font-bold text-(--color-text-primary) border border-(--color-border-muted) transition-colors hover:bg-(--color-border-muted) ${VOICE_FOCUS_RING_CLASS}`}
+                        className={`flex w-full items-center justify-center gap-2 rounded-2xl border border-(--color-border-muted) bg-(--color-surface-muted) py-3 font-bold text-(--color-text-primary) transition-colors hover:bg-(--color-border-muted) ${VOICE_FOCUS_RING_CLASS}`}
                     >
                         <Share2 size={18} aria-hidden="true" />
                         {shareLabel}
@@ -458,7 +486,7 @@ export function VoiceResultPanel({
                     {isSpeaking ? (
                         <button
                             onClick={onStopSpeaking}
-                            className={`flex w-full items-center justify-center gap-2 rounded-2xl bg-red-50 dark:bg-red-950/20 py-3 font-bold text-red-700 dark:text-red-400 transition-colors hover:bg-red-100 dark:hover:bg-red-900/30 ${VOICE_FOCUS_RING_CLASS}`}
+                            className={`flex w-full items-center justify-center gap-2 rounded-2xl bg-red-50 py-3 font-bold text-red-700 transition-colors hover:bg-red-100 dark:bg-red-950/20 dark:text-red-400 dark:hover:bg-red-900/30 ${VOICE_FOCUS_RING_CLASS}`}
                         >
                             <Square size={18} aria-hidden="true" />
                             {stopSpeakingLabel}
@@ -466,7 +494,7 @@ export function VoiceResultPanel({
                     ) : (
                         <button
                             onClick={onReplay}
-                            className={`flex w-full items-center justify-center gap-2 rounded-2xl bg-blue-50 dark:bg-blue-950/20 py-3 font-bold text-blue-700 dark:text-blue-450 transition-colors hover:bg-blue-100 dark:hover:bg-blue-900/30 ${VOICE_FOCUS_RING_CLASS}`}
+                            className={`dark:text-blue-450 flex w-full items-center justify-center gap-2 rounded-2xl bg-blue-50 py-3 font-bold text-blue-700 transition-colors hover:bg-blue-100 dark:bg-blue-950/20 dark:hover:bg-blue-900/30 ${VOICE_FOCUS_RING_CLASS}`}
                         >
                             <Play size={18} aria-hidden="true" />
                             {speakLabel}
@@ -476,7 +504,7 @@ export function VoiceResultPanel({
 
                 <button
                     onClick={onTryAgain}
-                    className={`flex w-full items-center justify-center gap-2 rounded-2xl bg-slate-900 dark:bg-slate-100 dark:text-slate-900 dark:hover:bg-slate-200 py-4 font-bold text-white transition-all hover:bg-slate-800 ${VOICE_FOCUS_RING_CLASS}`}
+                    className={`flex w-full items-center justify-center gap-2 rounded-2xl bg-slate-900 py-4 font-bold text-white transition-all hover:bg-slate-800 dark:bg-slate-100 dark:text-slate-900 dark:hover:bg-slate-200 ${VOICE_FOCUS_RING_CLASS}`}
                 >
                     <RotateCcw size={20} aria-hidden="true" />
                     {tryAgainLabel}
