+++ apps/web/components/ui/Skeleton.tsx
@@ -1,8 +1,6 @@
 import React from "react";
 
-export interface SkeletonProps extends React.HTMLAttributes<HTMLDivElement> {}
-
-export function Skeleton({ className = "", ...props }: SkeletonProps) {
+export function Skeleton({ className = "", ...props }: React.HTMLAttributes<HTMLDivElement>) {
     return (
         <div
             className={`relative overflow-hidden rounded-md bg-slate-200 before:absolute before:inset-0 before:-translate-x-full before:animate-[shimmer_2s_infinite] before:bg-gradient-to-r before:from-transparent before:via-white/20 before:to-transparent dark:bg-slate-800 ${className}`}
